mxshop/requirements.txt
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
decorator==4.3.0
diff-match-patch==20121119
Django==2.0.7
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-formtools==2.1
django-guardian==1.4.9
django-import-export==1.0.1
django-redis==4.9.0
django-reversion==3.0.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
DjangoUeditor==1.8.143
drf-extensions==0.3.1
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.11.3
idna==2.7
ipython==6.4.0
ipython-genutils==0.2.0
itypes==1.1.0
jdcal==1.4
jedi==0.12.1
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
mysqlclient==1.3.13
odfpy==1.3.6
openpyxl==2.5.4
parso==0.3.1
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
Pygments==2.2.0
PyJWT==1.6.4
pytz==2018.5
PyYAML==3.13
redis==2.10.6
requests==2.19.1
simplegeneric==0.8.1
six==1.11.0
tablib==0.12.1
traitlets==4.3.2
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.23
wcwidth==0.1.7
xlrd==1.1.0
XlsxWriter==1.0.5
xlwt==1.3.0
